requirements.txt
geopandas==1.0.1
numpy==1.26.4
pandas==2.2.2
pandas-datareader==0.10.0
pandas-gbq==0.26.1
pandas-stubs==2.2.2.240909
scikit-learn==1.6.1
sentence-transformers==3.3.1
sklearn-pandas==2.2.0
rapidfuzz==3.11.0